Move gen xml grid 040 tests
